Fix merge of individual umlaut structures into a single structure
Each umlaut file generates a umlaut structure. This structure has a
:nodes and a :diagrams key. In the beginning of the execution of a
umlaut script, we need to open, parse and merge all the individual
umlaut files into a single umlaut structure, which will be the input to
the generators. This merge was being done by a custom function, but it
could be simple replaced by the built-in merge function.